Moved tests to integration tests folder
