+++ jcli/tests/tests.rs
@@ -1,0 +1,4 @@
+extern crate jcli;
+
+pub mod common;
+pub mod transaction;
