+++ jcli/tests/transaction/mod.rs
@@ -1,0 +1,2 @@
+pub mod add_input;
+pub mod new;
